requirements.txt
joblib==1.4.2
matplotlib==3.9.2
numpy==2.1.3
pytorch_lightning==2.4.0
ruamel.base==1.0.0
scipy==1.14.1
torch==2.4.0
torchmetrics==1.4.1
tqdm==4.66.5
ruamel.yaml==0.17.40
opencv-python==4.10.0
